Add Atmel Studio version of bootloader
Adds an Atmel Studio project containing the example bootloader code. Due
to the lack of WinAVR's odd error, the startup sequence is slightly more
robust. In addtion, adds a working copy of example firmware. ALL FUTURE
FIRMWARE MUST DISABLE THE WDT IF NOT USED. This is critical for proper
firmware execution.